Moves ClustersDetailsService.create to ClustersService.create
